[refactor] community 어댑터 이름 변경하고 redis 폴더로 이동
