--- src/main/java/org/hana/wooahhanaapi/domain/community/adapter/ValidateAccountPort.java
@@ -1,7 +1,0 @@
-package org.hana.wooahhanaapi.domain.community.adapter;
-
-import org.hana.wooahhanaapi.domain.community.adapter.dto.AccountValidationConfirmDto;
-
-public interface ValidateAccountPort {
-    boolean validateAccount(AccountValidationConfirmDto dto);
-}
