+++ src/main/java/org/hana/wooahhanaapi/domain/community/controller/CommunityController.java
@@ -1,7 +1,7 @@
 package org.hana.wooahhanaapi.domain.community.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.hana.wooahhanaapi.domain.community.adapter.dto.SendValidationCodeReqDto;
+import org.hana.wooahhanaapi.utils.redis.dto.SendValidationCodeReqDto;
 import org.hana.wooahhanaapi.domain.community.dto.CommunityChgManagerReqDto;
 import org.hana.wooahhanaapi.domain.community.dto.CommunityCreateReqDto;
 import org.hana.wooahhanaapi.domain.community.dto.CommunityFeeStatusReqDto;
