+++ src/main/java/org/hana/wooahhanaapi/domain/community/service/CommunityService.java
@@ -5,11 +5,10 @@ import org.hana.wooahhanaapi.domain.account.adapter.AccountTransferPort;
 import org.hana.wooahhanaapi.domain.account.adapter.AccountTransferRecordPort;
 import org.hana.wooahhanaapi.domain.account.adapter.dto.*;
 import org.hana.wooahhanaapi.domain.account.exception.AccountNotFoundException;
-import org.hana.wooahhanaapi.domain.account.service.AccountService;
-import org.hana.wooahhanaapi.domain.community.adapter.ValidateAccountPort;
-import org.hana.wooahhanaapi.domain.community.adapter.dto.AccountValidationConfirmDto;
-import org.hana.wooahhanaapi.domain.community.adapter.SaveValidCodePort;
-import org.hana.wooahhanaapi.domain.community.adapter.dto.SendValidationCodeReqDto;
+import org.hana.wooahhanaapi.utils.redis.ValidateAccountPort;
+import org.hana.wooahhanaapi.utils.redis.dto.AccountValidationConfirmDto;
+import org.hana.wooahhanaapi.utils.redis.SaveValidCodePort;
+import org.hana.wooahhanaapi.utils.redis.dto.SendValidationCodeReqDto;
 import org.hana.wooahhanaapi.domain.account.exception.IncorrectValidationCodeException;
 import org.hana.wooahhanaapi.domain.community.domain.Community;
 import org.hana.wooahhanaapi.domain.community.dto.*;
