+++ src/main/java/org/hana/wooahhanaapi/utils/redis/RedisAdapter.java
@@ -1,9 +1,8 @@
-package org.hana.wooahhanaapi.domain.community.adapter;
+package org.hana.wooahhanaapi.utils.redis;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.hana.wooahhanaapi.domain.account.service.AccountService;
-import org.hana.wooahhanaapi.domain.community.adapter.dto.AccountValidationConfirmDto;
+import org.hana.wooahhanaapi.utils.redis.dto.AccountValidationConfirmDto;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Service;
@@ -14,7 +13,7 @@ import java.util.concurrent.TimeUnit;
 @Service
 @Slf4j
 @RequiredArgsConstructor
-public class CommunityAdapter implements ValidateAccountPort, SaveValidCodePort {
+public class RedisAdapter implements ValidateAccountPort, SaveValidCodePort {
     private final RedisTemplate<String, String> redisTemplate;
 
     @Override
