+++ src/main/java/org/hana/wooahhanaapi/utils/redis/SaveValidCodePort.java
@@ -1,4 +1,4 @@
-package org.hana.wooahhanaapi.domain.community.adapter;
+package org.hana.wooahhanaapi.utils.redis;
 
 public interface SaveValidCodePort {
     void saveValidCode(String accountNumber, String validCode);
